--- _infra/helm/action/templates/delete-cronjob.yaml
@@ -1,35 +1,0 @@
-apiVersion: batch/v1beta1
-kind: CronJob
-metadata:
-  name: {{ .Values.crons.recordDeleter.name }}
-spec:
-  schedule: "{{ .Values.crons.recordDeleter.cron }}"
-  jobTemplate:
-    spec:
-      template:
-        spec:
-          containers:
-          - name: {{ .Values.crons.recordDeleter.name }}
-            image: radial/busyboxplus:curl
-            env:
-            - name: SECURITY_USER_NAME
-              valueFrom:
-                secretKeyRef:
-                  name: security-credentials
-                  key: security-user
-            - name: SECURITY_USER_PASSWORD
-              valueFrom:
-                secretKeyRef:
-                  name: security-credentials
-                  key: security-password
-            - name: APP_NAME
-              value: {{ .Chart.Name }}
-            - name: APP_PORT
-              value: "{{ .Values.service.port }}"
-            - name: TARGET
-              value: {{ .Values.crons.recordDeleter.target }}
-            args:
-            - /bin/sh
-            - -c
-            - curl -X DELETE -u $(SECURITY_USER_NAME):$(SECURITY_USER_PASSWORD) http://$(APP_NAME):$(APP_PORT)/$(TARGET)
-          restartPolicy: OnFailure
