--- _infra/helm/action/templates/export-cronjob.yaml
@@ -1,31 +1,0 @@
-apiVersion: batch/v1beta1
-kind: CronJob
-metadata:
-  name: {{ .Values.crons.exportScheduler.name }}
-spec:
-  schedule: "{{ .Values.crons.exportScheduler.cron }}"
-  jobTemplate:
-    spec:
-      template:
-        spec:
-          containers:
-          - name: {{ .Values.crons.exportScheduler.name }}
-            image: radial/busyboxplus:curl
-            env:
-            - name: SECURITY_USER_NAME
-              valueFrom:
-                secretKeyRef:
-                  name: security-credentials
-                  key: security-user
-            - name: SECURITY_USER_PASSWORD
-              valueFrom:
-                secretKeyRef:
-                  name: security-credentials
-                  key: security-password
-            - name: TARGET
-              value: {{ .Values.crons.exportScheduler.target }}
-            args:
-            - /bin/sh
-            - -c
-            - curl -u $(SECURITY_USER_NAME):$(SECURITY_USER_PASSWORD) http://$(ACTION_SERVICE_HOST):$(ACTION_SERVICE_PORT)/$(TARGET)
-          restartPolicy: OnFailure
