+++ _infra/helm/action/values.yaml
@@ -42,14 +42,6 @@ crons:
     name: action-plan-executor
     cron: "* * * * *"
     target: "actionplans/execute"
-  recordDeleter:
-    name: action-export-old-record-deleter
-    cron: "0 3 * * *"
-    target: "delete-old-data"
-  exportScheduler:
-    name: action-export-scheduler
-    cron: "*/30 * * * *"
-    target: "export"
 
 resources:
   requests:
