--- src/main/java/uk/gov/ons/ctp/response/action/ActionExporterBeanMapper.java
@@ -1,38 +1,0 @@
-package uk.gov.ons.ctp.response.action;
-
-import ma.glasnost.orika.MapperFactory;
-import ma.glasnost.orika.converter.ConverterFactory;
-import ma.glasnost.orika.impl.ConfigurableMapper;
-import ma.glasnost.orika.impl.DefaultMapperFactory;
-import ma.glasnost.orika.impl.generator.EclipseJdtCompilerStrategy;
-import org.springframework.stereotype.Component;
-import uk.gov.ons.ctp.response.action.common.StringToUUIDConverter;
-import uk.gov.ons.ctp.response.action.domain.model.ActionRequestInstruction;
-import uk.gov.ons.ctp.response.action.representation.ActionRequestInstructionDTO;
-
-/** The bean mapper to go from Entity objects to Presentation objects. */
-@Component
-public class ActionExporterBeanMapper extends ConfigurableMapper {
-
-  @Override
-  public void configureFactoryBuilder(DefaultMapperFactory.Builder builder) {
-    builder.compilerStrategy(new EclipseJdtCompilerStrategy());
-  }
-
-  /**
-   * This method configures the bean mapper.
-   *
-   * @param factory the mapper factory
-   */
-  @Override
-  protected final void configure(final MapperFactory factory) {
-
-    ConverterFactory converterFactory = factory.getConverterFactory();
-    converterFactory.registerConverter(new StringToUUIDConverter());
-
-    factory
-        .classMap(ActionRequestInstruction.class, ActionRequestInstructionDTO.class)
-        .byDefault()
-        .register();
-  }
-}
