+++ src/main/java/uk/gov/ons/ctp/response/action/config/AppConfig.java
@@ -18,7 +18,6 @@ public class AppConfig {
   private CollectionExerciseSvc collectionExerciseSvc;
   private PartySvc partySvc;
   private SurveySvc surveySvc;
-  private DataGrid dataGrid;
   private ReportSettings reportSettings;
   private Rabbitmq rabbitmq;
   private Logging logging;
