--- src/main/java/uk/gov/ons/ctp/response/action/config/DataGrid.java
@@ -1,17 +1,0 @@
-package uk.gov.ons.ctp.response.action.config;
-
-import lombok.Data;
-import net.sourceforge.cobertura.CoverageIgnore;
-
-/** Config POJO for action plan exec params */
-@CoverageIgnore
-@Data
-public class DataGrid {
-  private String address;
-  private String password;
-  private Integer listTimeToWaitSeconds;
-  private Integer listTimeToLiveSeconds;
-  private Integer lockTimeToWaitSeconds;
-  private Integer lockTimeToLiveSeconds;
-  private Integer reportLockTimeToLiveSeconds;
-}
