+++ src/main/java/uk/gov/ons/ctp/response/action/service/NotificationFileCreator.java
@@ -2,11 +2,11 @@ package uk.gov.ons.ctp.response.action.service;
 
 import java.text.SimpleDateFormat;
 import java.time.Clock;
-import java.util.List;
+import java.util.*;
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
-import uk.gov.ons.ctp.response.action.domain.repository.ExportFileRepository;
 import uk.gov.ons.ctp.response.action.message.instruction.ActionRequest;
 
 @Service
@@ -17,20 +17,78 @@ public class NotificationFileCreator {
   private static final SimpleDateFormat FILENAME_DATE_FORMAT =
       new SimpleDateFormat("ddMMyyyy_HHmm");
 
-  private final ExportFileRepository exportFileRepository;
-
   private final Clock clock;
 
   private PrintFileService printFileService;
 
-  public NotificationFileCreator(
-      ExportFileRepository exportFileRepository, Clock clock, PrintFileService printFileService) {
-    this.exportFileRepository = exportFileRepository;
+  private class ExportData {
+    private List<ActionRequest> ariList;
+
+    public ExportData(ActionRequest ari) {
+      this.ariList = new ArrayList<>(Arrays.asList(ari));
+    }
+
+    public List<ActionRequest> getActionRequests() {
+      return ariList;
+    }
+
+    public void addActionRequest(ActionRequest ari) {
+      if (ariList != null) {
+        this.ariList.add(ari);
+      } else {
+        this.ariList = Arrays.asList(ari);
+      }
+    }
+  }
+
+  public NotificationFileCreator(Clock clock, PrintFileService printFileService) {
     this.clock = clock;
     this.printFileService = printFileService;
   }
 
-  public void uploadData(String filenamePrefix, List<ActionRequest> actionRequests) {
+  public void export(List<ActionRequest> actionRequests) {
+    Map<String, ExportData> filenamePrefixToDataMap = prepareData(actionRequests);
+
+    createAndSendFiles(filenamePrefixToDataMap);
+  }
+
+  private Map<String, ExportData> prepareData(List<ActionRequest> actionRequests) {
+    Map<String, ExportData> filenamePrefixToDataMap = new HashMap<>();
+
+    actionRequests.forEach(
+        ari -> {
+          String filenamePrefix =
+              FilenamePrefix.getPrefix(ari.getActionType())
+                  + "_"
+                  + ari.getSurveyRef()
+                  + "_"
+                  + getExerciseRefWithoutSurveyRef(ari.getExerciseRef());
+
+          if (filenamePrefixToDataMap.containsKey(filenamePrefix)) {
+            filenamePrefixToDataMap.get(filenamePrefix).addActionRequest(ari);
+          } else {
+            filenamePrefixToDataMap.put(filenamePrefix, new ExportData(ari));
+          }
+        });
+
+    return filenamePrefixToDataMap;
+  }
+
+  private void createAndSendFiles(Map<String, ExportData> filenamePrefixToDataMap) {
+
+    filenamePrefixToDataMap.forEach(
+        (filenamePrefix, data) -> {
+          List<ActionRequest> actionRequests = data.getActionRequests();
+          uploadData(filenamePrefix, actionRequests);
+        });
+  }
+
+  private String getExerciseRefWithoutSurveyRef(String exerciseRef) {
+    String exerciseRefWithoutSurveyRef = StringUtils.substringAfter(exerciseRef, "_");
+    return StringUtils.defaultIfEmpty(exerciseRefWithoutSurveyRef, exerciseRef);
+  }
+
+  private void uploadData(String filenamePrefix, List<ActionRequest> actionRequests) {
     if (actionRequests.isEmpty()) {
       log.info("no action request instructions to export");
       return;
@@ -38,15 +96,6 @@ public class NotificationFileCreator {
 
     final String now = FILENAME_DATE_FORMAT.format(clock.millis());
     String filename = String.format("%s_%s.csv", filenamePrefix, now);
-
-    if (exportFileRepository.existsByFilename(filename)) {
-      log.warn(
-          "filename: "
-              + filename
-              + ", duplicate filename. The cron job is probably running too frequently. The "
-              + "Action Exporter service is designed to only run every minute, maximum");
-      throw new RuntimeException();
-    }
 
     log.info("filename: " + filename + ", uploading file");
