+++ src/main/resources/database/changelog-master.yml
@@ -65,3 +65,6 @@ databaseChangeLog:
 
 - include:
     file: database/changes/release-21/changelog.yml
+
+- include:
+    file: database/changes/release-22/changelog.yml
