+++ src/main/resources/database/changes/release-22/changelog.yml
@@ -1,0 +1,12 @@
+databaseChangeLog:
+
+  - changeSet:
+      id: 22-1
+      author: Warren Bailey
+      changes:
+        - sqlFile:
+            comment: drop the action exporter tables
+            path: drop_action_exporter_tables.sql
+            relativeToChangelogFile: true
+            splitStatements: true
+
