+++ src/main/resources/database/changes/release-22/create_action_exporter_tables.sql
@@ -1,0 +1,5 @@
+DROP TABLE action.exportjob;
+DROP TABLE action.exportfile;
+DROP TABLE action.address;
+DROP TABLE action.contact;
+DROP TABLE action.actionrequest;
