+++ src/test/java/uk/gov/ons/ctp/response/action/message/ActionInstructionPublisherTest.java
@@ -19,7 +19,6 @@ import uk.gov.ons.ctp.response.action.message.instruction.Action;
 import uk.gov.ons.ctp.response.action.message.instruction.ActionCancel;
 import uk.gov.ons.ctp.response.action.message.instruction.ActionInstruction;
 import uk.gov.ons.ctp.response.action.message.instruction.ActionRequest;
-import uk.gov.ons.ctp.response.action.service.ActionExportService;
 import uk.gov.ons.ctp.response.lib.common.FixtureHelper;
 
 /** Tests for ActionInstructionPublisherImpl */
