+++ src/test/java/uk/gov/ons/ctp/response/action/service/ObjectBuilder.java
@@ -3,9 +3,6 @@ package uk.gov.ons.ctp.response.action.service;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
-import uk.gov.ons.ctp.response.action.domain.model.ActionRequestInstruction;
-import uk.gov.ons.ctp.response.action.domain.model.Address;
-import uk.gov.ons.ctp.response.action.domain.model.Contact;
 
 /** Utility class to build objects required in tests */
 public class ObjectBuilder {
